cloud_function/requirements.txt
PyPDF2==1.26.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-api-python-client==1.12.5
dacite==1.5.1
